optimize PUT foods[id], update swagger foods/[id]

+++ src/app/api/foods/[id]/route.ts
@@ -137,6 +137,36 @@
  *               image:
  *                 type: string
  *                 nullable: true
+ *               fibers_per_serving:
+ *                 type: number
+ *                 example: 10
+ *               sugars_per_serving:
+ *                 type: number
+ *                 example: 0
+ *               zincs_per_serving:
+ *                 type: number
+ *                 example: 0.003
+ *               magnesiums_per_serving:
+ *                 type: number
+ *                 example: 0.25
+ *               calciums_per_serving:
+ *                 type: number
+ *                 example: 0.1
+ *               irons_per_serving:
+ *                 type: number
+ *                 example: 0.003
+ *               vitamin_a_per_serving:
+ *                 type: number
+ *                 example: 0
+ *               vitamin_c_per_serving:
+ *                 type: number
+ *                 example: 0
+ *               vitamin_b12_per_serving:
+ *                 type: number
+ *                 example: 0
+ *               vitamin_d_per_serving:
+ *                 type: number
+ *                 example: 0
  *     responses:
  *       200:
  *         description: Food updated successfully
@@ -191,56 +221,74 @@
  *       401:
  *         description: Unauthorized
  */
-import { NextRequest } from 'next/server';
-import { FoodRepository } from '@/repositories/master.repository';
-import { foodSchema } from '@/validation/master.schema';
-import { successResponse, errorResponse } from '@/lib/response';
-import { fromZodError } from 'zod-validation-error';
-import { getAuthUser } from '@/lib/auth';
+import { NextRequest } from "next/server";
+import { FoodRepository } from "@/repositories/master.repository";
+import { foodSchema } from "@/validation/master.schema";
+import { successResponse, errorResponse } from "@/lib/response";
+import { fromZodError } from "zod-validation-error";
+import { getAuthUser } from "@/lib/auth";
 
-export async function GET(req: NextRequest, { params }: { params: Promise<{ id: string }> }) {
+export async function GET(
+  req: NextRequest,
+  { params }: { params: Promise<{ id: string }> },
+) {
   try {
     const { id } = await params;
     const foodId = parseInt(id);
     const food = await FoodRepository.findById(foodId);
-    if (!food) return errorResponse('Food not found', 404);
+    if (!food) return errorResponse("Food not found", 404);
     return successResponse(food);
   } catch (error: any) {
     return errorResponse(error.message, 500);
   }
 }
 
-export async function PUT(req: NextRequest, { params }: { params: Promise<{ id: string }> }) {
+export async function PUT(
+  req: NextRequest,
+  { params }: { params: Promise<{ id: string }> },
+) {
   try {
     const user = getAuthUser(req);
-    if (!user) return errorResponse('Unauthorized', 401);
+    if (!user) return errorResponse("Unauthorized", 401);
 
     const { id } = await params;
     const foodId = parseInt(id);
+
+    if ((await FoodRepository.findById(foodId)) === null) {
+      return errorResponse("Not exist food with that id", 404);
+    }
+
     const body = await req.json();
     const validation = foodSchema.safeParse(body);
     if (!validation.success) {
-      return errorResponse(fromZodError(validation.error).message);
+      return errorResponse("Wrong format",400);
     }
 
-    const food = await FoodRepository.update(foodId, validation.data);
-    if (!food) return errorResponse('Food not found', 404);
-    return successResponse(food, 'Food updated successfully');
+    const food = await FoodRepository.update(foodId,  validation.data);
+    if (!food) return errorResponse("Faild to update food", 404);
+    return successResponse(food, "Food updated successfully");
   } catch (error: any) {
     return errorResponse(error.message, 500);
   }
 }
 
-export async function DELETE(req: NextRequest, { params }: { params: Promise<{ id: string }> }) {
+export async function DELETE(
+  req: NextRequest,
+  { params }: { params: Promise<{ id: string }> },
+) {
   try {
     const user = getAuthUser(req);
-    if (!user) return errorResponse('Unauthorized', 401);
+    if (!user) return errorResponse("Unauthorized", 401);
 
     const { id } = await params;
     const foodId = parseInt(id);
+
+    if ((await FoodRepository.findById(foodId)) === null) {
+      return errorResponse("Not exist food with that id", 404);
+    }
+
     const deleted = await FoodRepository.delete(foodId);
-    if (!deleted) return errorResponse('Food not found', 404);
-    return successResponse(null, 'Food deleted successfully');
+    return successResponse(null, "Food deleted successfully");
   } catch (error: any) {
     return errorResponse(error.message, 500);
   }
